requirements.txt
imap_tools==1.7.4
gtts==2.5.4
playsound==1.3.0
python-dotenv==1.0.1
pygame==2.6.1